bif_type detection changed
Instead of comparing the rank of the full jacobian matrix with the rank of jac_red_jac_lam, branch points are now identified by changing sign of the determinante of the full Jacobian.
This is only testet for when a bifuraction is already detected.

Calculation of solver_jacobian_red/lam are only used when branch point is detected and therefore moved into the if-section.